Try to comply psr-4 standard
