+++ src/BlueskyConsts.php
@@ -1,0 +1,10 @@
+<?php
+
+class BlueskyConsts
+{
+    // don't change these unless Bluesky changes the limits
+    const MAX_UPLOAD_SIZE = 1000000;
+    const MAX_IMAGE_UPLOAD = 4;
+    const MIN_POST_SIZE = 3;
+    const MAX_POST_SIZE = 300;
+}
