+++ src/Version.php
@@ -1,0 +1,6 @@
+<?php
+
+class Version
+{
+    const VERSION = '2.0.11';
+}
