+++ src/php2BlueskyException.php
@@ -1,0 +1,3 @@
+<?php
+
+class php2BlueskyException extends \Exception {}
